fix compilation on 64bit machines, see JIRA GRB-27

+++ GRB/src/GRB/GRBSim.cxx
@@ -31,7 +31,7 @@ GRBSim::GRBSim(Parameters *params)
 void GRBSim::GetUniqueName(const void *ptr, std::string & name)
 {
   std::ostringstream my_name;
-  my_name << reinterpret_cast<int> (ptr);
+  my_name << reinterpret_cast<long> (ptr);
   name = my_name.str();
   gDirectory->Delete(name.c_str());
